requirements.txt
pandas
selenium
beautifulsoup4
html5lib
langdetect
pillow
wordcloud
matplotlib
pyspellchecker